Solve possible conflicts between `skipPat` and `lineBreakPat`
When there is an overlap between the strings the two patterns match,
in other words, a string S matches both patterns, should this string
be skipped or used as a line break? This possibility wasn't taken
into consideration in the old code, causing inconsistencies and
unnatural behavior.

The current solution is to give precedence to `lineBreakPat`: the
input text is first splitted using this pattern into lines, and then
`skipPat` is matched in each line.